requirements.txt
aiogram @ file:///home/ltshg/.cache/pypoetry/artifacts/01/9c/04/c1143501f88479532769c85a3081cb94699c9d4eba1079d3a0e341c8ea/aiogram-2.19-py3-none-any.whl
aiohttp @ file:///home/ltshg/.cache/pypoetry/artifacts/ec/c9/b8/045a900f6bd96c4ad567ba36e5cac7b7c00d1a7d0d53999cac47495ef0/aiohttp-3.8.1-cp310-cp310-manylinux_2_5_x86_64.manylinux1_x86_64.manylinux_2_12_x86_64.manylinux2010_x86_64.whl
aiosignal @ file:///home/ltshg/.cache/pypoetry/artifacts/30/30/0d/903be6c7ed45c2c0427a2a18486bb6ceb0a700c346a94414b6b395e16e/aiosignal-1.2.0-py3-none-any.whl
aiosqlite @ file:///home/ltshg/.cache/pypoetry/artifacts/60/a2/3a/29cf4aae59ac7c1a30ef747e8414e08e12af9d5f5be21d24055d093146/aiosqlite-0.17.0-py3-none-any.whl
async-timeout @ file:///home/ltshg/.cache/pypoetry/artifacts/5e/15/15/173c5bae6f8d28fa399857d021326b2b05fc2debcbc7bcae7c5bf79891/async_timeout-4.0.2-py3-none-any.whl
asyncpg @ file:///home/ltshg/.cache/pypoetry/artifacts/41/cd/7f/77ad6d6529fd3011699b3355b070ef4d72fb2d5a6eb1230125fe5817f0/asyncpg-0.25.0-cp310-cp310-manylinux_2_5_x86_64.manylinux1_x86_64.manylinux_2_12_x86_64.manylinux2010_x86_64.whl
attrs @ file:///home/ltshg/.cache/pypoetry/artifacts/c4/ee/75/fcd6c2f4c5a1d60b1bcd97014702ca01b3f032fd927d88780b38708409/attrs-21.4.0-py2.py3-none-any.whl
Babel @ file:///home/ltshg/.cache/pypoetry/artifacts/ed/4e/c7/8c4387255e163b817646635c2c88a7debca2361b2ae43c3a455fca786b/Babel-2.9.1-py2.py3-none-any.whl
certifi @ file:///home/ltshg/.cache/pypoetry/artifacts/89/06/5b/e4bfdf1b8370bdc0a26198e81afa98f68c01a8b1b515c9108c6ad98d8a/certifi-2021.10.8-py2.py3-none-any.whl
charset-normalizer @ file:///home/ltshg/.cache/pypoetry/artifacts/f4/80/c8/1b0ac4ebd0be6cfaa3e7257ab200f289221fe63d077c36855d2495dd11/charset_normalizer-2.0.11-py3-none-any.whl
frozenlist @ file:///home/ltshg/.cache/pypoetry/artifacts/ec/80/77/bdf92e2c93e0459cdcd039035f23def398b35a35e695631bef3abdb209/frozenlist-1.3.0-cp310-cp310-manylinux_2_5_x86_64.manylinux1_x86_64.manylinux_2_17_x86_64.manylinux2014_x86_64.whl
idna @ file:///home/ltshg/.cache/pypoetry/artifacts/90/36/8c/81eabf6ac88608721ab27f439c9a6b9a8e6a21cc58c59ebb1a42720199/idna-3.3-py3-none-any.whl
iso8601 @ file:///home/ltshg/.cache/pypoetry/artifacts/b5/5e/0f/27e34c5e5b6d45fe26ab25a5294f9aa5c54b7aa345012a404e567c8f65/iso8601-0.1.16-py2.py3-none-any.whl
loguru @ file:///home/ltshg/.cache/pypoetry/artifacts/36/1f/fa/e70e76ec49384e6de44a56eedd13608349891af6858a468e0f06d336a5/loguru-0.6.0-py3-none-any.whl
multidict @ file:///home/ltshg/.cache/pypoetry/artifacts/fc/03/de/7ff4d4eef94dd924e44bf6a7cd431a49013f48825f82687d4ba8bbb1e1/multidict-6.0.2-cp310-cp310-manylinux_2_17_x86_64.manylinux2014_x86_64.whl
numpy @ file:///home/ltshg/.cache/pypoetry/artifacts/e3/97/69/382ec5c4f63d1dbb6d75581ba246161e3e45e346797c68055279e1aa5e/numpy-1.22.2-cp310-cp310-manylinux_2_17_x86_64.manylinux2014_x86_64.whl
opencv-python @ file:///home/ltshg/.cache/pypoetry/artifacts/c6/f8/bb/0302d9a43105bd2389fe49ba30071ca7be24f41d52cfa9dbc88f50dcad/opencv_python-4.5.5.62-cp36-abi3-manylinux_2_17_x86_64.manylinux2014_x86_64.whl
Pillow @ file:///home/ltshg/.cache/pypoetry/artifacts/fb/b6/ed/2fc79fc3a4456bfcce72f6faea613c6febb3342ca041f3ac0244fed0f8/Pillow-9.0.1-cp310-cp310-manylinux_2_17_x86_64.manylinux2014_x86_64.whl
pypika-tortoise @ file:///home/ltshg/.cache/pypoetry/artifacts/dc/76/4d/58d551ee432fa8bf899ab2aff2640fca189d510d295f07c274250acbaa/pypika_tortoise-0.1.3-py3-none-any.whl
pytz @ file:///home/ltshg/.cache/pypoetry/artifacts/a0/f8/66/b50e6d372fa21e4d6db4b2fe5c5e94aaf59b82d02a7b3215fbc4fae889/pytz-2021.3-py2.py3-none-any.whl
tortoise-orm @ file:///home/ltshg/.cache/pypoetry/artifacts/db/1a/0c/336a15ffb13c2d09b6e03a86c3865be39d8aaa9dc56accbcdda9a983e0/tortoise_orm-0.18.1-py3-none-any.whl
typing_extensions @ file:///home/ltshg/.cache/pypoetry/artifacts/4f/0f/a7/b082eaf8890f4023be4c474d1c6eb672d28c6602d213a487c61e19a66b/typing_extensions-4.0.1-py3-none-any.whl
yarl @ file:///home/ltshg/.cache/pypoetry/artifacts/ab/46/71/ee32588139cc0f07f94e65cac6c281e6376e186842dbe17a31b0c4ea0f/yarl-1.7.2-cp310-cp310-manylinux_2_5_x86_64.manylinux1_x86_64.manylinux_2_12_x86_64.manylinux2010_x86_64.whl
